# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for SkillFort.",
        "title": "SkillFort-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "paths": {
...
        "/masterConfig": {
            "post": {
                "description": "Creates a new master config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "MasterConfig"
                ],
                "summary": "Create Master Config",
                "parameters": [
                    {
                        "description": "Master Config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.MasterConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/masterConfig/browse": {
            "get": {
                "description": "Get Master Config for browse",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "MasterConfig"
                ],
                "summary": "Get Master Config for browse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/masterConfig/value": {
            "get": {
                "description": "Get a specific master config by name",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "MasterConfig"
                ],
                "summary": "Get Master Config by Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Master Config Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/masterConfig/{id}": {
            "get": {
                "description": "Get a specific master config by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "MasterConfig"
                ],
                "summary": "Get Master Config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Master Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing master config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "MasterConfig"
                ],
                "summary": "Update Master Config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Master Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Master Config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.MasterConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Pagination": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "skipPagination": {
                    "type": "boolean"
                },
                "sort": {
                    "type": "string"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
...
        "errs.ErrorType": {
            "type": "string",
            "enum": [
                "OTHER",
                "INVALID",
                "IO",
                "EXIST",
                "NOT_EXIST",
                "PRIVATE",
                "INTERNAL",
                "BROKEN_LINK",
                "DATABASE",
                "VALIDATION",
                "UNANTICIPATED",
                "INVALID_REQUEST",
                "SMTP_ERROR",
                "EMAIL_ERROR",
                "STORAGE",
                "INSUFFICIENT_ACCESS",
                "CUSTOMER_DUPLICATE_TRANSACTION"
            ],
            "x-enum-comments": {
                "BROKEN_LINK": "Link target does not exist.",
                "CUSTOMER_DUPLICATE_TRANSACTION": "Duplicate Transaction",
                "DATABASE": "Error from database.",
                "EMAILERROR": "SMTP Error",
                "EXIST": "Item already exists.",
                "INSUFFICIENT_ACCESS": "Insufficient Access",
                "INTERNAL": "Internal error or inconsistency.",
                "INVALID": "Invalid operation for this type of item.",
                "INVALID_REQUEST": "Invalid Request",
                "IO": "External I/O error such as network failure.",
                "NOT_EXIST": "Item does not exist.",
                "OTHER": "Unclassified error. This value is not printed in the error message.",
                "PRIVATE": "Information withheld.",
                "SMTPERROR": "SMTP Error",
                "STORAGE": "Storage Error",
                "UNANTICIPATED": "Unanticipated error.",
                "VALIDATION": "Input validation error."
            },
            "x-enum-varnames": [
                "OTHER",
                "INVALID",
                "IO",
                "EXIST",
                "NOT_EXIST",
                "PRIVATE",
                "INTERNAL",
                "BROKEN_LINK",
                "DATABASE",
                "VALIDATION",
                "UNANTICIPATED",
                "INVALID_REQUEST",
                "SMTPERROR",
                "EMAILERROR",
                "STORAGE",
                "INSUFFICIENT_ACCESS",
                "CUSTOMER_DUPLICATE_TRANSACTION"
            ]
        },
        "errs.XError": {
            "type": "object",
            "properties": {
                "error": {},
                "errorCode": {
                    "type": "string"
                },
                "errorType": {
                    "$ref": "#/definitions/errs.ErrorType"
                },
                "params": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "requestModel.MasterConfig": {
            "type": "object",
            "properties": {
                "currentValue": {
                    "type": "string"
                },
                "defaultValue": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "previousValue": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "useDefault": {
                    "type": "boolean"
                }
            }
        },
...
        "response.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errs.XError"
                },
                "message": {
                    "type": "string"
                },
                "pageMetaData": {
                    "$ref": "#/definitions/db.Pagination"
                },
                "type": {
                    "$ref": "#/definitions/response.ResponseType"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errs.XError"
                },
                "message": {
                    "type": "string"
                },
                "pageMetaData": {
                    "$ref": "#/definitions/db.Pagination"
                },
                "type": {
                    "$ref": "#/definitions/response.ResponseType"
                }
            }
        },
        "response.ResponseType": {
            "type": "string",
            "enum": [
                "Success",
                "Failure",
                "Warning"
            ],
            "x-enum-varnames": [
                "SUCCESS",
                "FAILURE",
                "WARNING"
            ]
        },
...
    }
}